LPS-107036 Rename file
